renamed constraints and conditions
